requirements.txt
# AI-Powered Detailed Diagnostic Report (DDR) Generator
# Requirements file for Python dependencies

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# PDF Processing
PyMuPDF==1.23.26
pdfplumber==0.10.4

# AI Processing
google-generativeai==0.8.3

# PDF Report Generation
reportlab==4.1.0

# Image Processing (for image handling)
Pillow==10.2.0

# Logging and Utilities
python-dotenv==1.0.1

# Optional: For better text processing
nltk==3.8.1

# Development and Testing (optional)
pytest==8.1.1
pytest-cov==5.0.0
